Add UseCaseModule
