+++ app/src/main/java/seki/com/doyouworkout/di/ApplicationComponent.kt
@@ -8,12 +8,14 @@ import seki.com.doyouworkout.di.component.MainListActivityBuilder
 import seki.com.doyouworkout.di.component.NewWorkoutActivityBuilder
 import seki.com.doyouworkout.di.component.SettingActivityBuilder
 import seki.com.doyouworkout.ui.setting.SettingActivity
+import seki.com.doyouworkout.usecase.UseCaseModule
 import javax.inject.Singleton
 
 @Singleton
 @Component(modules = [
     AndroidSupportInjectionModule::class,
     ApplicationModule::class,
+    UseCaseModule::class,
     WorkoutViewModelModule::class,
     MainListActivityBuilder::class,
     NewWorkoutActivityBuilder::class,
