+++ app/src/main/java/seki/com/doyouworkout/di/ApplicationModule.kt
@@ -11,16 +11,12 @@ import seki.com.doyouworkout.data.ResourceSupplierImp
 import seki.com.doyouworkout.data.cache.Cache
 import seki.com.doyouworkout.data.cache.DataCache
 import seki.com.doyouworkout.data.db.AppDataBase
-import seki.com.doyouworkout.data.db.mapper.WorkoutMapper
 import seki.com.doyouworkout.data.repository.LocalRepository
 import seki.com.doyouworkout.data.repository.LocalRepositoryImp
 import seki.com.doyouworkout.data.repository.Repository
 import seki.com.doyouworkout.data.repository.WorkoutRepository
-import seki.com.doyouworkout.usecase.*
-import seki.com.doyouworkout.usecase.impl.FetchTrainingUseCaseImp
-import seki.com.doyouworkout.usecase.impl.GetWorkoutUseCaseImp
-import seki.com.doyouworkout.usecase.impl.UpdateTrainingUseCaseImp
-import seki.com.doyouworkout.usecase.impl.UpdateWorkoutUseCaseImp
+import seki.com.doyouworkout.usecase.SchedulersProvider
+import seki.com.doyouworkout.usecase.SchedulersProviderBase
 import javax.inject.Singleton
 
 @Module
@@ -68,38 +64,5 @@ class ApplicationModule(private val applicationContext: Context) {
     @Provides
     fun provideSchedulerProvider(): SchedulersProviderBase {
         return SchedulersProvider()
-    }
-
-    @Singleton
-    @Provides
-    fun provideGetWorkoutUseCase(
-            repository: WorkoutRepository,
-            mapper: WorkoutMapper
-    ): GetWorkoutUseCase {
-        return GetWorkoutUseCaseImp(repository, mapper)
-    }
-
-    @Singleton
-    @Provides
-    fun provideUpdateWorkoutUseCase(
-            repository: WorkoutRepository
-    ): UpdateWorkoutUseCase {
-        return UpdateWorkoutUseCaseImp(repository)
-    }
-
-    @Singleton
-    @Provides
-    fun provideFetchTrainingUseCase(
-            repository: WorkoutRepository
-    ): FetchTrainingUseCase {
-        return FetchTrainingUseCaseImp(repository)
-    }
-
-    @Singleton
-    @Provides
-    fun provideUpdateTrainingUseCase(
-            repository: WorkoutRepository
-    ): UpdateTrainingUseCase {
-        return UpdateTrainingUseCaseImp(repository)
     }
 }
